переносит функции рендера в отдельный файл
