+++ src/utils.js
@@ -1,21 +1,3 @@
-export const RenderPosition = {
-  AFTERBEGIN: `afterbegin`,
-  BEFOREEND: `beforeend`,
-  AFTEREND: `afterend`,
-};
-
-
-export const render = (container, element, place) => {
-  switch (place) {
-    case RenderPosition.AFTERBEGIN:
-      container.prepend(element);
-      break;
-    case RenderPosition.BEFOREEND:
-      container.append(element);
-      break;
-  }
-};
-
 export const getRandomInteger = (min, max) => {
   // случайное число от min до (max+1)
   const rand = min + Math.random() * (max + 1 - min);
@@ -83,11 +65,4 @@ export const createSlug = (str) => {
     .replace(/-+/g, `-`);
 
   return str;
-};
-
-export const createElement = (template) => {
-  const newElement = document.createElement(`div`);
-  newElement.innerHTML = template;
-
-  return newElement.firstElementChild;
 };
